Check service finalizer on upgrade test and fix-up test cases
